+++ Content.Server/Imperial/Medieval/Horn/MedievalHornPlaytimeRequirementComponent.cs
@@ -1,0 +1,6 @@
+namespace Content.Server.Imperial.Medieval.Horn;
+
+[RegisterComponent]
+public sealed partial class MedievalHornPlaytimeRequirementComponent : Component
+{
+}
