+++ Resources/Prototypes/Imperial/Medieval/medieval.yml
@@ -4031,6 +4031,7 @@
   placement:
     mode: SnapgridCenter
   components:
+  - type: MedievalHornPlaytimeRequirement
   - type: MeleeSound
     soundGroups:
       Brute:
